# api/swagger/docs/swagger.yaml
basePath: /
definitions:
  common.SwaggerErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      data: {}
      message:
        example: invalid_request
        type: string
    type: object
  common.SwaggerResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: ok
        type: string
    type: object
...
  v1.CategoryResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/v1.CategoryResponse'
        type: array
      created_at:
        type: string
      document_count:
        type: integer
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      slug:
        type: string
      sort:
        type: integer
      updated_at:
        type: string
    type: object
...
  v1.CreateCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      sort:
        type: integer
    type: object
...
  v1.CreateTagRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
  v1.ListCategoriesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.CategoryResponse'
        type: array
    type: object
...
  v1.ListTagsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.TagResponse'
        type: array
    type: object
...
  v1.TagResponse:
    properties:
      created_at:
        type: string
      document_count:
        type: integer
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  v1.UpdateCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      sort:
        type: integer
    type: object
...
  v1.UpdateTagRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
...
info:
  contact: {}
  description: 'REST API for Knowledge Core (v1). JSON envelope: code (HTTP-style),
    message, data. 中文：Knowledge Core REST API（v1）；响应为 code、message、data 的 JSON 包装。'
  title: Knowledge Core HTTP API
  version: "1.0"
paths:
  /api/v1/admin/categories:
    get:
      description: List categories as a tree. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListCategoriesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Admin Categories
    post:
      consumes:
      - application/json
      description: Create a category. Admin only.
      parameters:
      - description: Category create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Admin Categories
  /api/v1/admin/categories/{id}:
    delete:
      description: Delete an empty leaf category. Admin only.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Admin Categories
    patch:
      consumes:
      - application/json
      description: Update a category. Admin only.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Admin Categories
...
  /api/v1/admin/tags:
    get:
      description: List tags ordered by usage. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListTagsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List tags
      tags:
      - Admin Tags
    post:
      consumes:
      - application/json
      description: Create a tag. Admin only.
      parameters:
      - description: Tag create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.TagResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - Admin Tags
  /api/v1/admin/tags/{id}:
    delete:
      description: Delete a tag and clear document tag links. Admin only.
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - Admin Tags
    patch:
      consumes:
      - application/json
      description: Update a tag. Admin only.
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.TagResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - Admin Tags
...
  /api/v1/categories:
    get:
      description: List categories as a tree.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListCategoriesResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      summary: List categories
      tags:
      - Categories
...
  /api/v1/tags:
    get:
      description: List tags ordered by usage.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListTagsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      summary: List tags
      tags:
      - Tags
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use the format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"